Moving test utilities from __init__ to separate module
