+++ satpy/tests/multiscene_tests/test_blend.py
@@ -27,7 +27,7 @@ import xarray as xr
 from pyresample.geometry import AreaDefinition
 
 from satpy import DataQuery
-from satpy.tests.multiscene_tests import _create_test_area, _create_test_dataset, _create_test_int8_dataset
+from satpy.tests.multiscene_tests.test_utils import _create_test_area, _create_test_dataset, _create_test_int8_dataset
 from satpy.tests.utils import make_dataid
